pyproject.toml
[project]
name = "homework-hub"
version = "0.1.0"
description = "Aggregates homework from Google Classroom, Compass and Edrolo into per-child Google Sheets."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Ryan Northam" }]
dependencies = [
    "apscheduler>=3.10.4",
    "click>=8.1.7",
    "fastapi>=0.115.0",
    "google-api-python-client>=2.149.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "gspread>=6.1.4",
    "httpx>=0.27.2",
    "playwright>=1.48.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "structlog>=24.4.0",
    "uvicorn>=0.32.0",
]

[project.scripts]
homework-hub = "homework_hub.__main__:cli"

[dependency-groups]
dev = [
    "black>=24.10.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "responses>=0.25.3",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/homework_hub"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # black handles line length

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
